requirements.txt
streamlit
langchain-openai
langgraph
python-dotenv
requests
pytest
pytest-dotenv
langchain-groq
python-docx
fpdf2
markdown2
